test: move mockWarn into setup files
