+++ packages/runtime-core/__tests__/apiInject.spec.ts
@@ -10,13 +10,9 @@ import {
   reactive
 } from '../src/index'
 import { render, nodeOps, serialize } from '@vue/runtime-test'
-import { mockWarn } from '@vue/shared'
 
 // reference: https://vue-composition-api-rfc.netlify.com/api.html#provide-inject
-
 describe('api: provide/inject', () => {
-  mockWarn()
-
   it('string keys', () => {
     const Provider = {
       setup() {
