+++ packages/runtime-core/__tests__/rendererAttrsFallthrough.spec.ts
@@ -13,11 +13,8 @@ import {
   createCommentVNode,
   Fragment
 } from '@vue/runtime-dom'
-import { mockWarn } from '@vue/shared'
 
 describe('attribute fallthrough', () => {
-  mockWarn()
-
   it('should allow attrs to fallthrough', async () => {
     const click = jest.fn()
     const childUpdated = jest.fn()
